ai-models/requirements.txt
# Core ML/DL frameworks
tensorflow>=2.13.0
keras>=2.13.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.2
